+++ text_recognition/deep-text-recognition-benchmark/deep_text_recognition_benchmark_utils.py
